requirements.txt
requests 
pandas 
duckdb 
scikit-learn 
umap-learn 
fastapi 
uvicorn 
python-dotenv 
matplotlib
